Svelte: Refactor: Move storiesOf examples under "Legacy" prefix
