+++ examples/svelte-kitchen-sink/src/stories/legacy/addon-actions.stories.js
@@ -1,10 +1,10 @@
 import { action } from '@storybook/addon-actions';
 
 import ButtonView from './views/ButtonView.svelte';
-import Button from '../components/Button.svelte';
+import Button from '../../components/Button.svelte';
 
 export default {
-  title: 'Addon|Actions',
+  title: 'Legacy|Addon/Actions',
 };
 
 export const actionOnViewMethod = () => ({
