+++ examples/svelte-kitchen-sink/src/stories/legacy/addon-backgrounds.stories.js
@@ -1,7 +1,7 @@
 import ButtonView from './views/ButtonView.svelte';
 
 export default {
-  title: 'Addon|Backgrounds',
+  title: 'Legacy|Addon/Backgrounds',
   parameters: {
     backgrounds: [
       { name: 'light', value: '#eeeeee' },
