+++ examples/svelte-kitchen-sink/src/stories/legacy/addon-knobs.stories.js
@@ -3,7 +3,7 @@ import { withKnobs, text, number } from '@storybook/addon-knobs';
 import ActionKnobView from './views/ActionKnobView.svelte';
 
 export default {
-  title: 'Addon|Knobs',
+  title: 'Legacy|Addon/Knobs',
   decorators: [withKnobs],
 };
