+++ examples/svelte-kitchen-sink/src/stories/legacy/button.stories.js
@@ -1,7 +1,7 @@
 import ButtonView from './views/ButtonView.svelte';
 
 export default {
-  title: 'Button',
+  title: 'Legacy|Button',
 };
 
 export const rounded = () => ({
