+++ examples/svelte-kitchen-sink/src/stories/legacy/views/ActionLinkView.svelte
