+++ examples/svelte-kitchen-sink/src/stories/legacy/views/WelcomeView.svelte
